Extract CRUDService and stub out types
